Simplied tx sending, readded toasts
