+++ src/actions/changeLiquidity.ts
@@ -7,7 +7,7 @@ import {
   PERPETUALS_ADDRESS,
   TRANSFER_AUTHORITY,
 } from "@/utils/constants";
-import { automaticSendTransaction } from "@/utils/dispatchTransaction";
+import { automaticSendTransaction } from "@/utils/TransactionHandlers";
 import { createAtaIfNeeded, wrapSolIfNeeded } from "@/utils/transactionHelpers";
 import { BN } from "@project-serum/anchor";
 import { getAssociatedTokenAddress, TOKEN_PROGRAM_ID } from "@solana/spl-token";
@@ -126,7 +126,10 @@ export async function changeLiquidity(
     methodBuilder = methodBuilder.preInstructions(preInstructions);
 
   try {
-    await automaticSendTransaction(methodBuilder);
+    await automaticSendTransaction(
+      methodBuilder,
+      perpetual_program.provider.connection
+    );
   } catch (err) {
     console.log(err);
     throw err;
