+++ src/components/Notify.tsx
@@ -1,7 +1,12 @@
-import { toast } from 'react-toastify';
+import { toast } from "react-toastify";
 
 // @ts-ignore
-export const notify = (text: string | JSX.Element, type = "success", hideProgressBar=false) => toast[type](text, {
+export const notify = (
+  text: string | JSX.Element,
+  type = "success",
+  hideProgressBar = false
+) =>
+  toast[type](text, {
     position: "bottom-left",
     autoClose: 3000,
     hideProgressBar: hideProgressBar,
@@ -9,4 +14,4 @@ export const notify = (text: string | JSX.Element, type = "success", hideProgres
     pauseOnHover: true,
     draggable: true,
     progress: undefined,
-});
+  });
